+++ client/src/utils/hooks.js
@@ -19,15 +19,26 @@ function useClientInfos() {
     // Try to parse the cookie into a JavaScript object
     clientInfos = cookies['client-infos'].split(',');
 
-    if(clientInfos.length !== 3) {
+    if(clientInfos.length <= 3) {
       window.location.href = '/cosmos-ui/logout';
     }
     
-    return {
+    let res = {
       nickname: clientInfos[0],
       userRole: clientInfos[1],
       role: clientInfos[2]
     };
+
+    if(clientInfos.length > 3) {
+      let roleUntil = new Date(parseInt(clientInfos[3], 10) * 1000);
+      let currentDate = new Date();
+
+      if(roleUntil < currentDate && res.userRole == "2") {
+        res.userRole = "1";
+      }
+    }
+
+    return res;
   } catch (error) {
     console.error('Error parsing client-infos cookie:', error);
     return {