+++ src/docker/export.go
@@ -41,7 +41,12 @@ func ExportContainer(containerID string) (ContainerCreateRequestContainer, error
 			User:         detailedInfo.Config.User,
 			Tty:          detailedInfo.Config.Tty,
 			StdinOpen:    detailedInfo.Config.OpenStdin,
-			Hostname:     detailedInfo.Config.Hostname,
+			Hostname:     func () string { 
+				if string(detailedInfo.HostConfig.NetworkMode) == "bridge" || string(detailedInfo.HostConfig.NetworkMode) == "default" {
+					return detailedInfo.Config.Hostname
+				}
+				return ""
+			}(),
 			Domainname:   detailedInfo.Config.Domainname,
 			MacAddress:   detailedInfo.NetworkSettings.MacAddress,
 			NetworkMode:  string(detailedInfo.HostConfig.NetworkMode),