+++ changelog.md
@@ -1,3 +1,14 @@
+## Version 0.19.0
+ - Constellation allows nodes to see and ping each others
+ - Constellation now has a firewall!
+ - Constellation now has exit nodes
+ - Improve docker image cleanup efficiency
+ - Improve support for container network modes in import/export
+ - Fixed the annoying "user unauthenticated" error when opening the homepage after the admin token expired
+ - Fixed issue with exporting hostname when it would be incompatible to re-importing it
+ - Updating network mode now also updates the network-mode label
+ - Default storage path is now /cosmos-storage instead of /usr
+
 ## Version 0.18.4
  - Fix issue with DB credentials dissapearing
  - Remove expired discount